Dave Perry:
This patch makes the pa24 changes to use the mag.sw from Instruments 3d.
Changed the mag_switch.ac to include the pick rectangles and the
mags.xml to use the pick automation.  Also includes the edits to
pa28-161.xml and pa28-161-set.xml to use the 3d radio stack.